add cmd.c / add print_struct_debug func
